backend/requirements.txt
asgiref==3.9.1
cffi==1.17.1
cryptography==45.0.5
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
joblib==1.5.1
numpy==2.2.6
opencv-python==4.12.0.88
pandas==2.3.1
pillow==11.3.0
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.1
scipy==1.16.0
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
